got checkout working not wonky wohooo
